read() returns a signed value, use int instead of size_t
git-svn-id: https://projects.maemo.org/svn/af/packages/fremantle/trunk/hildon-desktop@41033 7feff2eb-67fc-0310-aff0-fef983b50a43

+++ src/launcher/hd-app-mgr.c
@@ -1401,7 +1401,7 @@ hd_app_mgr_check_loadavg (void)
   if (fd >= 0)
     {
       char buffer[32];
-      size_t size = read (fd, buffer, sizeof(buffer) -1);
+      int size = read (fd, buffer, sizeof(buffer) -1);
 
       close (fd);
       if (size > 0)
